reshuffle HorizontalCardListFeatureConnected files to match existing "ConnectedComponent" paradigm.
- fixes stories import error for this component too
