+++ packages/apollos-ui-connected/src/HorizontalCardListFeatureConnected/HorizontalCardListFeature/HorizontalCardListFeature.tests.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import { Providers } from '../utils/testUtils';
+import { Providers } from '../../utils/testUtils';
 
-import HorizontalCardListFeature from './HorizontalCardListFeature';
+import HorizontalCardListFeature from '.';
 
 const cards = [
   {
